requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0

# LLM API Client
openai>=1.12.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Vector Database
chromadb>=0.4.22

# Embedding API
dashscope>=1.14.0

# Utilities
python-multipart>=0.0.6
